September 7'th
First session back to developing plugin after my trip hiatus. I meant to return sooner but the fear of starting back up again kept me at bay.
this session focused on moving forward and thinking through the current issues. (I may have not mentioned this but MUCH of my work with this plugin development has been through the assistance of ChatGPT while I am learning all of this stuff for the first time. Much is new, so having ChatGPT there as a way to bounce ideas and to give examples helps a lot)
I worried about simultaneous command usages and how UUID's are saved via methods within each of the commands. And the topic came to maybe switching away from the YML file format and using a database like SQLite.
So this session involved gutting out all of the YML config info, setting up a database, and setting up all of CRUD for use in the database.

While not entirely successful as I have issues with the inventory data being serialized, it was a MASSIVE step forward.